refactor(components): turned all class in function
