+++ src/App.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import './App.css';
-import NavBar from './components/NavBar/NavBar';
+
+import { NavBar } from './components/index';
 
 function App() {
   return (
