+++ src/components/Rectangular-Btn/Rectangular-Btn.tsx
@@ -1,23 +1,19 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
+import React from 'react';
 import './Rectangular-Btn.css';
 
+import { Link } from 'react-router-dom';
 import { Button } from 'antd';
 
 type Props = {
   children: string;
 };
 
-class RectangularBtn extends Component<Props> {
-  render() {
-    return (
-      <Link to="#">
-        <Button type="primary" className="rectangular-btn">
-          {this.props.children}
-        </Button>
-      </Link>
-    );
-  }
+export function RectangularBtn(props: Props) {
+  return (
+    <Link to="#">
+      <Button type="primary" className="rectangular-btn">
+        {props.children}
+      </Button>
+    </Link>
+  );
 }
-
-export default RectangularBtn;
