+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { CreateProjectFormModal } from './CreateProjectFormModal/CreateProjectFormModal';
+export { NavBar } from './NavBar/NavBar';
+export { NavBarApp } from './NavBarApp/NavBarApp';
+export { ProjectCard } from './ProjectCard/ProjectCard';
+export { RectangularBtn } from './Rectangular-Btn/Rectangular-Btn';
