+++ src/pages/Project/Project.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './Project.css';
 
 function Project() {
   return (
